requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
colorama>=0.4.6
